minor cleanup: fix types

+++ crates/api/src/domains/art/service.rs
@@ -32,7 +32,7 @@ impl ARTService {
     pub async fn get_art(
         &self,
         chat_id: &Uuid,
-        sequence_number: Option<i64>,
+        sequence_number: Option<u32>,
     ) -> Result<ARTRecord<ARTGroup>, ARTServiceError> {
         let arts_storage = MongoARTStorage::new().await?;
         let record = match &sequence_number {
@@ -52,7 +52,7 @@ impl ARTService {
     pub async fn get_previous_art(
         &self,
         chat_id: &Uuid,
-        sequence_number: Option<i64>,
+        sequence_number: Option<u32>,
     ) -> Result<ARTRecord<ARTGroup>, ARTServiceError> {
         info!(
             "Retreiving previous art for sequence_number: {}",
@@ -93,7 +93,7 @@ impl ARTService {
     pub async fn get_art_by_sequence_number(
         &self,
         chat_id: &Uuid,
-        sequence_number: i64,
+        sequence_number: u32,
     ) -> Result<ARTRecord<ARTGroup>, ARTServiceError> {
         let art_record = self.get_initial_art(chat_id).await?;
         let mut initial_art = art_record.art;
@@ -178,8 +178,8 @@ impl ARTService {
         &self,
         chat_id: &Uuid,
         filter: Document,
-        limit: i64,
-        skip: i64,
+        limit: u32,
+        skip: u32,
     ) -> Result<Vec<ARTChangesRecord<ARTGroup>>, ARTServiceError> {
         let record = MongoARTChangesStorage::new(chat_id)
             .await?
